[Feat] #140 - AlertViewBuildable BaseFeature로 이동 및 분리
