+++ SOPT-iOS/Projects/Domain/Sources/Model/StampGuideDataModel.swift
