+++ SOPT-iOS/Projects/Features/StampFeature/Sources/ListDetailScene/VC/ListDetailVC.swift
@@ -20,6 +20,7 @@ import DSKit
 
 import Lottie
 
+import BaseFeatureDependency
 import StampFeatureInterface
 
 public enum TextViewState {
@@ -77,7 +78,7 @@ public class ListDetailVC: UIViewController, ListDetailViewControllable {
     // MARK: - Properties
     
     public var viewModel: ListDetailViewModel!
-    public var factory: StampFeatureViewBuildable!
+    public var factory: (AlertViewBuildable & StampFeatureViewBuildable)!
     private var cancelBag = CancelBag()
     private var sceneType: ListDetailSceneType {
         get {
